Add assets recovered from hard drive
For some reason these images were not backed up along with the code.
I'm a moron so I reinstalled windows over the drive these files were on.
I managed to use R-Studio to save most of the game assets. Metadata on
these images was lost, including file names, so I had to parse through a
lot of images found by R-Studio to find these, which are correct as far
as I remember. File names were made to match the code.

FlyingEnemyGIF.gif could not be found and will be forever missed.