Less classes extend Thread
Initial work to untangle the Threads. This sped up certain aspects of the game that I did not intend to, not sure why. It could be that running way less threads is revealing that the sleep counts I was using did not correspond to the experience of the game.

Back when I first wrote this, Threads was
the only way I could figure out how to make
everything move. Ha.